% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_Valencia_clusters.R
\name{assign_to_Valencia_clusters}
\alias{assign_to_Valencia_clusters}
\title{Assign samples to Valencia clusters}
\usage{
assign_to_Valencia_clusters(input, distance = "YC")
}
\arguments{
\item{input}{a `matrix` or `data.frame` where each row is a sample for
which (sub)CST needs to be assigned}

\item{distance}{a `character` specifying which metric should be used to
measure (dis)similarity between sample composition and Valencia clusters}
}
\value{
a `list` with 3 components
- `distances`: the distance matrix between each sample (rows of `input`) and
Valencia clusters
- `assignment`: a `data.frame` with the CST and subCST assignment for each
sample (rows of `input`)
- `missing_taxa`: a `character` with the list of taxa in Valencia clusters
that are missing in the `input`.
}
\description{
Assign samples to Valencia clusters
}
