% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_Valencia_clusters.R
\name{BC}
\alias{BC}
\title{Computes the Bray-Curtis dissimilarity between the rows of two matrices}
\usage{
BC(mat1, mat2 = NULL)
}
\arguments{
\item{mat1}{the matrix with the first set of samples}

\item{mat2}{(optional) the matric with the second set of samples}
}
\value{
a distance matrix whose dimensions are `nrow(mat1)` x `nrow(mat2)`
}
\description{
Computes the Bray-Curtis dissimilarity between the rows of two matrices
}
