% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Valencia_tax_table.R
\name{build_Valencia_tax_table}
\alias{build_Valencia_tax_table}
\title{Builds a pseudo taxonomic table from Valencia taxa names}
\usage{
build_Valencia_tax_table(taxa_vec)
}
\arguments{
\item{taxa_vec}{a \code{character} vector with the names of the taxa as formatted by Valencia}
}
\value{
a \code{data.frame} with the pseudo taxonomic table for the Valencia
taxonomic labels in \code{taxa_vec}
}
\description{
Builds a pseudo taxonomic table from Valencia taxa names
}
