% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_Valencia_taxonomy.R
\name{convert_to_Valencia_taxonomy}
\alias{convert_to_Valencia_taxonomy}
\title{"Converts" a taxonomic count/proportion table to match the Valencia cluster taxonomy.}
\usage{
convert_to_Valencia_taxonomy(input, tax_table)
}
\arguments{
\item{input}{a table of ASV/OTU/species counts or proportions.}

\item{tax_table}{the taxonomic table associated to the columns of \code{input}.}
}
\value{
a \code{list} with two elements. The first element is the
\code{conversion_table}, which is the \code{tax_table} augmented with the
Valencia label for each columns of \code{input}.
The second element is the \code{converted_input}: the counts/proportions from
the original \code{input} aggregated for each Valencia taxonomic label.
}
\description{
"Converts" a taxonomic count/proportion table to match the Valencia cluster taxonomy.
}
