% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_Valencia_taxonomy.R
\name{.convert}
\alias{.convert}
\title{Aggregate the counts/proportions from the input matrix by taxonomic label}
\usage{
.convert(input, conversion_table)
}
\arguments{
\item{input}{a `matrix` with counts and proportions (samples are rows, species are columns)}

\item{conversion_table}{a `data.frame` with at least two columns:}
}
\value{
a `matrix`
}
\description{
Aggregate the counts/proportions from the input matrix by taxonomic label
}
