% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_topics.R
\name{.label_topics_1_model}
\alias{.label_topics_1_model}
\title{Label topics from a single model to match the closest Valencia sub-CST}
\usage{
.label_topics_1_model(lda_model, tax_table, distance)
}
\arguments{
\item{lda_model}{a single LDA model (output of `topicmodels::LDA`).}

\item{tax_table}{the taxonomic table for each row of the `beta` matrix}

\item{distance}{the distance used to compute the (dis)similarity between the
topic composition and the Valencia centroids.}
}
\value{
a single LDA model where the rownames of `beta` and the colnames
of `gamma` match the label of the closest Valencia sub-CST
}
\description{
Label topics from a single model to match the closest Valencia sub-CST
}
