% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset}
\format{
## `example_data`
A `phyloseq` object with the Vaginal samples from https://purl.stanford.edu/yb681vm1809
}
\usage{
example_data
}
\description{
Example dataset
}
\keyword{datasets}
