% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valencia_clusters.R
\name{get_Valencia_clusters}
\alias{get_Valencia_clusters}
\title{Retrieve Valencia centroids}
\usage{
get_Valencia_clusters()
}
\value{
\code{matrix} composition of clusters (rows are subCST, columns are taxa)
}
\description{
Retrieve Valencia centroids
}
\examples{
get_Valencia_clusters()

}
