% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_topics.R
\name{label_topics}
\alias{label_topics}
\title{Label topics to match the closest Valencia sub-CST}
\usage{
label_topics(lda_models, tax_table, distance)
}
\arguments{
\item{lda_models}{a single LDA model (output of `topicmodels::LDA`) or
a list of LDA models (e.g., output of `alto::fit_lda_models`).}

\item{tax_table}{the taxonomic table for each row of the `beta` matrix}

\item{distance}{the distance used to compute the (dis)similarity between the
topic composition and the Valencia centroids.}
}
\value{
a single LDA model or a list of LDA models (depending on the input)
where the rownames of `beta` and the colnames of `gamma` match the label
of the closest Valencia sub-CST
}
\description{
Label topics to match the closest Valencia sub-CST
}
