% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_Valencia_taxonomy.R
\name{match_taxonomies}
\alias{match_taxonomies}
\title{Matches taxonomic tables}
\usage{
match_taxonomies(tax_table, v_tax_table)
}
\arguments{
\item{tax_table}{a \code{data.frame} with the original taxonomic assignments}

\item{v_tax_table}{a \code{data.frame} with the}
}
\value{
a \code{data.frame} containing the "conversion table" between the
original taxa and the Valencia taxonomic labels.
}
\description{
Matches taxonomic tables
}
