% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomic_levels}
\alias{taxonomic_levels}
\title{Taxonomic levels}
\format{
## `taxonomic_levels`
A `vector` with the ordered taxonomic levels (from higher to lower order)
}
\usage{
taxonomic_levels
}
\description{
Ordered list of taxonomic levels required in taxonomic tables
}
\keyword{datasets}
